conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2024.11.26=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.4.0=pypi_0
google-api-core=2.23.0=pypi_0
google-api-python-client=2.154.0=pypi_0
google-auth=2.36.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.1=pypi_0
googleapis-common-protos=1.66.0=pypi_0
httplib2=0.22.0=pypi_0
idna=3.10=pypi_0
joblib=1.4.2=pypi_0
libffi=3.4.4=hd77b12b_1
numpy=2.1.3=pypi_0
oauthlib=3.2.2=pypi_0
openssl=3.0.15=h827c3e9_0
pandas=2.2.3=pypi_0
pip=24.2=py311haa95532_0
proto-plus=1.25.0=pypi_0
protobuf=5.29.0=pypi_0
pyarrow=18.1.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pyparsing=3.2.0=pypi_0
python=3.11.10=h4607a30_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h0416ee5_0
tzdata=2024.2=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.3=pypi_0
vc=14.40=h2eaa2aa_1
vs2015_runtime=14.40.33807=h98bb1dd_1
wheel=0.44.0=py311haa95532_0
xz=5.4.6=h8cc25b3_1
zlib=1.2.13=h8cc25b3_1
